ft-313 added github logo and updated readme
