+++ client/src/assets/imgs/GitHub-Mark/PNG/GitHub-Mark-Light-32px.png:Zone.Identifier
@@ -1,0 +1,3 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=G:\Projects\cover\abc_practice\GitHub-Mark.zip
