+++ client/src/components/Logo/index.tsx
@@ -3,14 +3,12 @@ import React from "react";
 import { Div } from "atomize";
 import "./style.scss";
 
-const Logo = ({ height = "20" }: { width?: string; height?: string }): JSX.Element => {
+const Logo = ({ className = "" }: { className?: string }): JSX.Element => {
   return (
-    <Div d="flex" p="0.7rem" align="center">
-      <img
-        height={height}
-        src="https://github.com/Amijakan/mochatto/blob/main/logo/logo_light.png?raw=true"
-      />
-    </Div>
+    <img
+    className={className}
+      src="https://github.com/Amijakan/mochatto/blob/main/logo/logo_light.png?raw=true"
+    />
   );
 };
