+++ client/src/templates/RoomTemplate/style.scss
@@ -1,0 +1,3 @@
+.logo-img {
+  height: 20px;
+}
